+++ tedi/components/buttons/index.ts
@@ -1,4 +1,6 @@
 export * from "./button/button.component";
+export * from "./button-group";
+export * from "./card-button";
 export * from "./info-button/info-button.component";
 export * from "./collapse/collapse.component";
 export * from "./collapse-button/collapse-button.component";