+++ tedi/components/content/index.ts
@@ -1,3 +1,4 @@
+export * from "./card";
 export * from "./list/list.component";
 export * from "./text-group/";
 export * from "./carousel";